refactor: removed state sync context due to uselessness
