+++ frontend/src/__tests__/integration/api/auth.test.tsx
@@ -21,19 +21,6 @@ jest.mock("next/navigation", () => ({
   })),
 }));
 
-// Context 모킹 - 실제 구현에 맞게 정확하게 모킹
-jest.mock("@/contexts/state/StateSyncContext", () => ({
-  useStateSync: jest.fn(() => ({
-    publish: jest.fn(),
-    subscribe: jest.fn(() => jest.fn()),
-    getState: jest.fn(() => ({})),
-    syncStates: jest.fn(),
-    clearState: jest.fn(),
-    clearAllStates: jest.fn(),
-  })),
-  StateSyncProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
-}));
-
 jest.mock("@/contexts/navigation/NavigationContext", () => ({
   useNavigation: jest.fn(() => ({
     activeTab: 0,
@@ -460,14 +447,6 @@ jest.mock("@/contexts/AppContext", () => ({
     session: {
       data: null,
       status: "unauthenticated"
-    },
-    stateSync: {
-      publish: jest.fn(),
-      subscribe: jest.fn(() => jest.fn()),
-      getState: jest.fn(() => ({})),
-      syncStates: jest.fn(),
-      clearState: jest.fn(),
-      clearAllStates: jest.fn(),
     },
     goBack: jest.fn(),
     updateForm: jest.fn(),
